fixed bug with sega in parsing of extension of file and created Makefile

+++ src/parse_op.c
@@ -29,7 +29,7 @@ int             find_oper(char *str, int len)
     while (i < 16)
     {
         if (ft_strnequ(g_op_tab[i].name, str, len) &&
-                ft_strlen(g_op_tab[i].name) == len)
+		        (int)ft_strlen(g_op_tab[i].name) == len)
             return (g_op_tab[i].op_code);
         i++;
     }
